move all the validator version info code to the browser side component
