+++ .eslintrc.json
@@ -9,10 +9,12 @@
     "jsx-a11y/click-events-have-key-events": "off",
     "jsx-a11y/no-static-element-interactions": "off"
   },
-  "extends": ["erb"],
+  "extends": [
+    "erb"
+  ],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "ecmaVersion": 2021,
+    "ecmaVersion": 2022,
     "ecmaFeatures": {
       "jsx": true
     },
@@ -26,5 +28,8 @@
       "typescript": {}
     }
   },
-  "plugins": ["react", "@typescript-eslint"]
+  "plugins": [
+    "react",
+    "@typescript-eslint"
+  ]
 }
